refactor(commands): split command helper modules
